refactor(interpreter): extract oracle pattern matching into eval/patterns.rs
Move evaluate_oracle, evaluate_oracle_branch, type_of_scrutinee, the
three destructuring matchers (scroll / artifact / lexicon), and
values_match_for_pattern out of eval/statements.rs into a dedicated
eval/patterns.rs, together with the seven oracle-focused unit tests
(including both scope-leak regressions). evaluate_oracle stays the
single pub(super) entry point; no behaviour or public-API change.

Closes #507
